Extract key metrics settings card to component.
