+++ assets/js/components/settings/SettingsAdmin.js
@@ -19,119 +19,27 @@
 /**
  * WordPress dependencies
  */
-import { useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
-import { addQueryArgs } from '@wordpress/url';
 
 /**
  * Internal dependencies
  */
-import Data from 'googlesitekit-data';
-import { CORE_USER } from '../../googlesitekit/datastore/user/constants';
-import { CORE_SITE } from '../../googlesitekit/datastore/site/constants';
-import { CORE_LOCATION } from '../../googlesitekit/datastore/location/constants';
 import Layout from '../layout/Layout';
 import { Grid, Cell, Row } from '../../material-components';
 import OptIn from '../OptIn';
 import ResetButton from '../ResetButton';
-import UserInputPreview from '../user-input/UserInputPreview';
-import { USER_INPUT_QUESTIONS_LIST } from '../user-input/util/constants';
 import { useFeature } from '../../hooks/useFeature';
-import { trackEvent } from '../../util';
+import SettingsCardKeyMetrics from './SettingsCardKeyMetrics';
 import SettingsPlugin from './SettingsPlugin';
-import useViewContext from '../../hooks/useViewContext';
-import SettingsKeyMetrics from './SettingsKeyMetrics';
-import Link from '../Link';
-const { useSelect, useDispatch } = Data;
 
 export default function SettingsAdmin() {
-	const viewContext = useViewContext();
 	const userInputEnabled = useFeature( 'userInput' );
-	const isUserInputCompleted = useSelect(
-		( select ) =>
-			userInputEnabled && select( CORE_USER ).isUserInputCompleted()
-	);
-	const userInputURL = useSelect( ( select ) =>
-		select( CORE_SITE ).getAdminURL( 'googlesitekit-user-input' )
-	);
-
-	const { navigateTo } = useDispatch( CORE_LOCATION );
-	const goTo = ( questionIndex = 1 ) => {
-		const questionSlug = USER_INPUT_QUESTIONS_LIST[ questionIndex - 1 ];
-		if ( questionSlug ) {
-			trackEvent( viewContext, 'question_edit', questionSlug );
-
-			navigateTo(
-				addQueryArgs( userInputURL, {
-					question: questionSlug,
-					redirect_url: global.location.href,
-					single: 'settings', // Allows the user to edit a single question then return to the settings page.
-				} )
-			);
-		}
-	};
-
-	const hasUserPickedMetrics = useSelect( ( select ) =>
-		select( CORE_USER ).getUserPickedMetrics()
-	);
-	const ctaLabel = !! hasUserPickedMetrics?.length
-		? __( 'Set your site goals', 'google-site-kit' )
-		: __( 'Personalize your metrics', 'google-site-kit' );
-
-	useEffect( () => {
-		if ( isUserInputCompleted ) {
-			trackEvent( viewContext, 'summary_view' );
-		}
-	}, [ isUserInputCompleted, viewContext ] );
 
 	return (
 		<Row>
 			{ userInputEnabled && (
 				<Cell size={ 12 }>
-					<Layout
-						title={ __( 'Key metrics', 'google-site-kit' ) }
-						header
-						rounded
-					>
-						<div className="googlesitekit-settings-module googlesitekit-settings-module--active googlesitekit-settings-user-input">
-							<SettingsKeyMetrics />
-							<Grid>
-								{ isUserInputCompleted && (
-									<Row>
-										<Cell size={ 12 }>
-											<UserInputPreview
-												goTo={ goTo }
-												noHeader
-												noFooter
-												settingsView
-												showIndividualCTAs
-											/>
-										</Cell>
-									</Row>
-								) }
-								{ isUserInputCompleted === false && (
-									<Row>
-										<Cell
-											className="googlesitekit-user-input__notification"
-											size={ 12 }
-										>
-											<p>
-												<span>
-													{ __(
-														'Answer 3 quick questions to help us show the most relevant data for your site',
-														'google-site-kit'
-													) }
-												</span>
-											</p>
-											<Link href={ userInputURL }>
-												{ ctaLabel }
-											</Link>
-										</Cell>
-									</Row>
-								) }
-							</Grid>
-						</div>
-					</Layout>
+					<SettingsCardKeyMetrics />
 				</Cell>
 			) }
